requirements.txt

Flask==2.3.2
Flask-WTF>=1.1.1
bcrypt==4.0.1
psycopg2-binary==2.9.6
python-dotenv==1.0.0
pandas==2.0.3
Werkzeug>=2.3.3
numpy==1.24.3
boxsdk
gunicorn==21.2.0 